Restructure ObjectFileReader/Writer for improved debugging
