+++ TactilityCore/CMakeLists.txt
@@ -9,6 +9,7 @@ if (DEFINED ENV{ESP_IDF_VERSION})
     idf_component_register(
         SRCS ${SOURCE_FILES}
         INCLUDE_DIRS "Include/"
+        PRIV_INCLUDE_DIRS "Private/"
         REQUIRES mbedtls nvs_flash esp_rom esp_timer
     )
 
@@ -22,6 +23,10 @@ else()
 
     target_sources(TactilityCore
         PRIVATE ${SOURCES}
+    )
+
+    include_directories(
+        PRIVATE Private/
     )
 
     target_include_directories(TactilityCore SYSTEM
